Added bare-metal documents.
Part 2 of 5: blueprint general-bare-metal-provisioning-framework.

This documents describe how to install and configure bare-metal nova-compute
to support pxe and non-pxe(tilera) back-end machines.
It includes the packages needed for installation,
how to create a bare-metal instance type,
how to register bare-metal hosts and NIC, and how to run an instance.
Nova flags for the configuration of bare-metal nova-compute are described.

Change-Id: I6358e5ede094b4d887d15913b19f877169a0f99c
Co-authored-by: Mikyung Kang <mkkang@isi.edu>
Co-authored-by: David Kang <dkang@isi.edu>
Co-authored-by: Ken Igarashi <igarashik@nttdocomo.co.jp>
Co-authored-by: Arata Notsu <notsu@virtualtech.jp>