Added bare-metal host manager.
Part 4 of 5: blueprint general-bare-metal-provisioning-framework.

The bare-metal host manager supports bare-metal provisioning and
original VM provisioning. For VM provisioning, it simply calls
the original host manager. A bare-metal nova-compute may manage
multiple bare-metal machines and it's capability report is not
it's own capability but the capability of the bare-metal
machines being managed. The existing 'capability' is designed
to describe the capability of one machine, and cannot describe
the full capabilities of the multiple bare-metal machines.
So, the existing 'capability' is used to show the maximum
resource in the bare-metal machine farm.
The bare-metal host manager detects if a nova-compute is
for VM or bare-metal. For bare-metal machine provisioning,
it fetches the full capabilities of a bare-metal nova-compute
by accessing bare-metal database.
The full capabilities are to be used by the scheduler.
No changes are needed in the scheduler side.

Change-Id: Ib207f495493db99a4a6cd1ac0044662fb7397322
Co-authored-by: Mikyung Kang <mkkang@isi.edu>
Co-authored-by: David Kang <dkang@isi.edu>
Co-authored-by: Ken Igarashi <igarashik@nttdocomo.co.jp>
Co-authored-by: Arata Notsu <notsu@virtualtech.jp>