Added separate bare-metal MySQL DB.
Part 1 of 5: blueprint general-bare-metal-provisioning-framework

Previously, bare-metal provisioning used text files to store information of
bare-metal machines. In this patch, a MySQL database is used to store the
information. The database is designed to support pxe/non-pee
booting methods, heterogeneous hypervisor types, and architectures.
Using a MySQL database makes maintenance and upgrades easier than
using text files.

The database for bare-metal machines is implemented as a separate database
from the main nova database. The database can be on any machine.
The location of the database and its server need to be specified as a flag
in the nova.conf file (as in the case of glance).

There are a couple of reasons for this approach.
First, the information needed for bare-metal machines is different from that
for non-bare-metal machines. With a separate database
for bare-metal machines, the database can be customized without affecting
the main nova database.

Second, fault-tolerance can be embedded in bare-metal nova-compute.
Since a single instance of bare-metal nova-compute manages multiple bare-metal
nodes, fault-tolerance of a bare-metal nova-compute node is very important.
With a separate database for bare-metal machines, fault-tolerance can be
achieved independently from the main nova database.

Replication of the bare-metal database and implementation of fault-tolerance
are not part of this patch.

The implementation models nova and its database as much as possible.

The bare-metal driver must be upgraded to use this database.
The patches of the bare-metal driver will be proposed
after this patch is merged upstream.

Change-Id: I7b7ba1903a672a50c567f95fc6554d119463b0c5
Co-authored-by: Mikyung Kang <mkkang@isi.edu>
Co-authored-by: David Kang <dkang@isi.edu>
Co-authored-by: Ken Igarashi <igarashik@nttdocomo.co.jp>
Co-authored-by: Arata Notsu <notsu@virtualtech.jp>